Implement redux
